refactor: 쓸모없는 Slack 라우터 제거
