+++ src/routes/index.ts
@@ -5,7 +5,6 @@ import NotiRouter from './noti.router';
 import LeaderboardRouter from './leaderboard.router';
 import TotalStatsRouter from './totalStats.router';
 import WebhookRouter from './webhook.router';
-import SlackRouter from './slack.router';
 
 
 const router: Router = express.Router();
@@ -20,7 +19,5 @@ router.use('/', NotiRouter);
 router.use('/', LeaderboardRouter);
 router.use('/', TotalStatsRouter);
 router.use('/', WebhookRouter);
-router.use('/', SlackRouter);
-
 
 export default router;
